Remove need for common parameters.
