+++ scripts/templateHelpers.test.ts
@@ -14,11 +14,8 @@ import {
   getObjectIdByAssetId,
   getParameterTypes,
   isAllowedTrait,
-  isCommonPathParameter,
-  isCommonQueryParameter,
   loud,
 } from './templateHelpers';
-import { commonParameterPositions } from '../src/static/commonParameters';
 import { ASSET_OBJECT_MAP } from './config';
 
 /** Technically, `type` could be much more, but for tests this will do. AMF is tricky. */
@@ -161,55 +158,6 @@ describe('Test formatForTsDoc template help function', () => {
     expect(formatForTsDoc('\nthis is a spaced     line')).toStrictEqual(
       '\nthis is a spaced     line',
     );
-  });
-});
-
-describe('Test isCommonPathParameter template help function', () => {
-  it('returns false for null input', () => {
-    expect(isCommonPathParameter(null)).toBe(false);
-  });
-
-  it('returns false for empty string', () => {
-    expect(isCommonPathParameter('')).toBe(false);
-  });
-
-  it('returns false for not a common parameter', () => {
-    expect(isCommonPathParameter('not-a-common-parameter')).toBe(false);
-  });
-
-  it('returns true for a common parameter', () => {
-    expect(isCommonPathParameter(commonParameterPositions.pathParameters[0])).toBe(true);
-  });
-
-  it('returns true for all common parameter', () => {
-    commonParameterPositions.pathParameters.forEach((p) => {
-      expect(isCommonPathParameter(p)).toBe(true);
-    });
-  });
-});
-
-describe('Test isCommonQueryParameter template help function', () => {
-  it('returns false for null input', () => {
-    expect(isCommonQueryParameter(null)).toBe(false);
-  });
-
-  it('returns false for empty string', () => {
-    expect(isCommonQueryParameter('')).toBe(false);
-  });
-
-  it('returns false for not a common parameter', () => {
-    expect(isCommonQueryParameter('not-a-common-parameter')).toBe(false);
-  });
-
-  it('returns true for a common parameter', () => {
-    expect(isCommonQueryParameter(commonParameterPositions.queryParameters[0]))
-      .toBe(true);
-  });
-
-  it('returns true for all common parameter', () => {
-    commonParameterPositions.queryParameters.forEach((p) => {
-      expect(isCommonQueryParameter(p)).toBe(true);
-    });
   });
 });
