+++ src/static/helpers.ts
@@ -11,3 +11,31 @@ export type CompositeParameters<MethodParameters extends object, ConfigParameter
 
 export type RequireParametersUnlessAllAreOptional<T extends { parameters?: object }> =
     {} extends NonNullable<T['parameters']> ? T : T & Required<Pick<T, 'parameters'>>;
+
+/**
+ * Template parameters used in the base URI of all API endpoints. `version` will default to `"v1"`
+ * if not specified.
+ */
+export interface BaseUriParameters {
+    shortCode: string;
+    version?: string; // Optional, will default to "v1" if not provided.
+  }
+
+/**
+   * Generic interface for path parameters.
+   */
+export interface PathParameters {
+    [key: string]: string | number | boolean;
+  }
+
+/**
+   * Generic interface for query parameters.
+   */
+export interface QueryParameters {
+    [key: string]: string | number | boolean | string[] | number[];
+  }
+
+/**
+   * Generic interface for all parameter types.
+   */
+export type UrlParameters = PathParameters | QueryParameters;
