+++ templates/client.ts.hbs
@@ -1,8 +1,11 @@
 import { fetch, Request }  from "cross-fetch";
 
 import ClientConfig, { ClientConfigInit } from "./clientConfig";
-import type { CommonBaseUriParameters, CommonParameters } from "./commonParameters";
-import type { CompositeParameters, RequireParametersUnlessAllAreOptional } from "./helpers";
+import type {
+  BaseUriParameters,
+  CompositeParameters,
+  RequireParametersUnlessAllAreOptional
+} from "./helpers";
 import TemplateURL from "./templateUrl";
 import { USER_AGENT_HEADER, USER_AGENT_VALUE } from "./version";
 
@@ -48,9 +51,9 @@ export type {{name.upperCamelCase}}QueryParameters = {
 }
 
 /**
- * All common parameters and all parameters that are used by {{{name.upperCamelCase}}}.
+ * All parameters that are used by {{{name.upperCamelCase}}}.
  */
-export type {{name.upperCamelCase}}Parameters = {{name.upperCamelCase}}PathParameters & {{name.upperCamelCase}}QueryParameters;
+export type {{name.upperCamelCase}}Parameters = {{name.upperCamelCase}}PathParameters & BaseUriParameters & {{name.upperCamelCase}}QueryParameters;
 
 /**
 * [{{{metadata.name}}}](https://developer.commercecloud.com/s/api-details/{{{getObjectIdByAssetId metadata.assetId}}})
@@ -84,7 +87,7 @@ export type {{name.upperCamelCase}}Parameters = {{name.upperCamelCase}}PathParam
 * 
 
 */
-export class {{name.upperCamelCase}}<ConfigParameters extends {{{name.upperCamelCase}}}Parameters & CommonParameters> {
+export class {{name.upperCamelCase}}<ConfigParameters extends {{{name.upperCamelCase}}}Parameters & Record<string, unknown>> {
   public clientConfig: ClientConfig<ConfigParameters>;
 
   static readonly defaultBaseUri = "{{getBaseUriFromDocument model}}";
