+++ templates/index.ts.hbs
@@ -2,7 +2,6 @@
 export { {{name.upperCamelCase}} } from "./{{name.lowerCamelCase}}";
 {{/each}}
 export type { ClientConfigInit } from "./clientConfig";
-export type { CommonParameters } from "./commonParameters";
 // Can't currently use `export from` with default exports
 import ClientConfig from "./clientConfig";
 import TemplateURL from "./templateUrl";
