+++ templates/operations.ts.hbs
@@ -119,7 +119,7 @@
       const optionParams = options?.parameters || ({} as Partial<NonNullable<NonNullable<typeof options>["parameters"]>>);
       const configParams = this.clientConfig.parameters;
 
-      const pathParams: {{{@root.name.upperCamelCase}}}PathParameters & Required<CommonBaseUriParameters> = {
+      const pathParams: {{{@root.name.upperCamelCase}}}PathParameters & Required<BaseUriParameters> = {
         {{!-- Base URI parameters are well known and unlikely to change, so they are hard-coded for simplicity  --}}
         shortCode: configParams.shortCode,
         version: configParams.version || "v1"
@@ -138,7 +138,6 @@
       {{/each}}
 
       const queryParams: {{{@root.name.upperCamelCase}}}QueryParameters = {};
-      // Fall back to config for common query parameters
       {{#each request.queryParameters}}
       if (optionParams["{{{name}}}"] !== undefined) {
         queryParams["{{{name}}}"] = optionParams["{{{name}}}"];
